changing right formation for error message on cd function and debugging echo
