+++ srcs/builtins/ft_echo.c
@@ -6,7 +6,7 @@
 /*   By: llitovuo <llitovuo@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/24 10:05:18 by llitovuo          #+#    #+#             */
-/*   Updated: 2024/06/04 14:48:00 by llitovuo         ###   ########.fr       */
+/*   Updated: 2024/06/11 09:55:09 by llitovuo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
